[MOD] Account_analytic_analysis: Avoid expected singleton.
When opening project/analytic_account from sale_order an error stating expected singleton
would be raised because it was possible that our sale_order would be picked in multiple
moves, and invoiced like this. The result would be that 1 sale_order_line would have
multiple account_invoice_lines linked to it. Therefore we chose to extend the list and
always look over the linked_invoice_lines' ids